Konverterer modals.jsx til tsx for å fikse log spamming
